requirements.txt
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
drf-yasg==1.20.0
psycopg2-binary==2.9.3
Pillow==9.0.0
gunicorn==20.1.0
uritemplate==3.0.1
PyYAML==5.4.1
humanfriendly==10.0
django-polymorphic==3.1.0
django-filter==21.1
